Python_Solution/requirements.txt
mysql_connector_repackaged==0.3.1
pika==1.3.1
requests==2.28.1
requests_mock==1.10.0
